refactor: extract recommendation hooks modules
