+++ frontend/src/hooks/recommendations/controller.ts
@@ -1,0 +1,114 @@
+import { FormEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react'
+
+import type { Region } from '../../types'
+import { RecommendationRow, buildRecommendationRows, formatWeekLabel } from '../../utils/recommendations'
+
+import { useRecommendationLoader } from './loader'
+import { useCropCatalog } from '../useCropCatalog'
+
+export interface UseRecommendationsOptions {
+  favorites: readonly number[]
+  initialRegion?: Region
+}
+
+export interface UseRecommendationsResult {
+  region: Region
+  setRegion: (region: Region) => void
+  queryWeek: string
+  setQueryWeek: (week: string) => void
+  currentWeek: string
+  displayWeek: string
+  sortedRows: RecommendationRow[]
+  handleSubmit: (event: FormEvent<HTMLFormElement>) => void
+}
+
+export const useRecommendations = ({ favorites, initialRegion }: UseRecommendationsOptions): UseRecommendationsResult => {
+  const initialRegionRef = useRef<Region>(initialRegion ?? 'temperate')
+  const [region, setRegion] = useState<Region>(initialRegionRef.current)
+  const regionSyncRef = useRef<Region>(initialRegionRef.current)
+  const regionFetchSkipRef = useRef<Region | null>(null)
+  const { catalog: cropCatalog } = useCropCatalog()
+  const cropIndex = useMemo(() => {
+    const map = new Map<string, { id: number; category?: string }>()
+    cropCatalog.forEach((entry, cropName) => {
+      map.set(cropName, { id: entry.id, category: entry.category })
+    })
+    return map
+  }, [cropCatalog])
+  const { queryWeek, setQueryWeek: setRawQueryWeek, activeWeek, items, currentWeek, requestRecommendations } =
+    useRecommendationLoader(region)
+
+  const setQueryWeek = useCallback(
+    (nextWeek: string) => {
+      setRawQueryWeek(nextWeek)
+    },
+    [setRawQueryWeek],
+  )
+
+  useEffect(() => {
+    if (initialRegion !== undefined && initialRegion !== initialRegionRef.current) {
+      initialRegionRef.current = initialRegion
+      setRegion(initialRegion)
+    }
+  }, [initialRegion, setRegion])
+
+  useEffect(() => {
+    if (regionSyncRef.current === region) {
+      return
+    }
+    regionSyncRef.current = region
+    if (regionFetchSkipRef.current === region) {
+      regionFetchSkipRef.current = null
+      return
+    }
+    void requestRecommendations(currentWeek, { regionOverride: region })
+  }, [currentWeek, region, requestRecommendations])
+
+  const sortedRows = useMemo<RecommendationRow[]>(() => {
+    return buildRecommendationRows({ items, favorites, cropIndex })
+  }, [items, cropIndex, favorites])
+
+  const handleSubmit = useCallback(
+    (event: FormEvent<HTMLFormElement>) => {
+      event.preventDefault()
+      const form = event.currentTarget
+      const weekField = form.elements.namedItem('week') as HTMLInputElement | null
+      const regionField = form.elements.namedItem('region') as HTMLSelectElement | null
+      const submittedWeek = weekField?.value ?? queryWeek
+      const submittedRegion = (regionField?.value as Region | undefined) ?? region
+      if (submittedRegion && submittedRegion !== region) {
+        setRegion(submittedRegion)
+      }
+      const regionChanged = submittedRegion !== undefined && submittedRegion !== region
+      const targetRegion = submittedRegion ?? region
+      const shouldRequest = !regionChanged || submittedWeek !== currentWeek
+      if (regionChanged && shouldRequest) {
+        regionFetchSkipRef.current = submittedRegion
+      }
+      if (shouldRequest) {
+        void requestRecommendations(submittedWeek, { regionOverride: targetRegion })
+      }
+    },
+    [currentWeek, queryWeek, region, requestRecommendations, setRegion],
+  )
+
+  const displayWeek = useMemo(() => formatWeekLabel(activeWeek), [activeWeek])
+
+  return {
+    region,
+    setRegion,
+    queryWeek,
+    setQueryWeek,
+    currentWeek,
+    displayWeek,
+    sortedRows,
+    handleSubmit,
+  }
+}
+
+export type { RecommendationRow } from '../../utils/recommendations'
+export type { RecommendationFetcher } from '../recommendationFetcher'
+export { useCropCatalog } from '../useCropCatalog'
+export type { CropCatalogEntry, CropCatalogMap, UseCropCatalogResult } from '../useCropCatalog'
+export { useRecommendationLoader } from './loader'
+export type { UseRecommendationLoaderResult } from './loader'
